[Subject*]: 2.0.0重大升级 [Type*   ]: feat [Scope   ]: [Reviewer]: [Body*   ]: 优化代码结构             数值类型由double改为decimal             参数支持boolean和datetime类型             支持常用数学函数             支持常用字符串函数             支持牛顿法一元函数计算             支持换行，支持#号注释             支持参数属性获取方法             ExpressionSetting增强             无参数时使用常用默认值计算，数值默认0，字符串默认空字符串，可通过ExpressionSetting设置             其它优化支持

+++ HLC.Expression/Expressions/ConstantExpression.cs
@@ -29,7 +29,7 @@ namespace HLC.Expression
 
             if (Type == ExpressionType.NumberConstant)
             {
-                return Result((double) Value);
+                return Result((decimal) Value);
             }
 
             if (Type == ExpressionType.RangeConstant)
@@ -59,7 +59,6 @@ namespace HLC.Expression
 
                 return sb.ToString();
             }
-
             return $"{Value}";
         }
     }
